+++ modules/oip042/artifact.go
@@ -136,7 +136,7 @@ func on42JsonEditArtifact(any jsoniter.Any, tx *datastore.TransactionData, sig s
 		Block:        tx.Block,
 		BlockHash:    tx.BlockHash,
 		Completed:    false,
-		Defective:    false,
+		Invalid:      false,
 		Signature:    sig,
 		Time:         tx.Transaction.Time,
 		Tx:           tx,